mvp-dev/requirements.txt
# MVP 开发环境 - Python 依赖
# 核心框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Celery 异步任务
celery==5.3.6
redis==5.0.1
flower==2.0.1

# 向量数据库客户端
qdrant-client==1.7.0

# MLflow 实验追踪
mlflow==2.9.2

# API 集成
requests==2.31.0
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1
slack-sdk==3.27.0
PyGithub==2.1.1

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0

# 工具库
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
